Common header for transport namespace
